# src/document/swagger-final.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sekolah Kopi Raisa",
    "version": "1.0.0",
    "description": "API untuk aplikasi Sekolah Kopi Raisa - E-commerce dan Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://be-web-sekolah-kopi-raisa.vercel.app",
      "description": "Vercel Production Server"
    },
    {
      "url": "https://sekolah-kopi-raisa.up.railway.app",
      "description": "Railway Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  },
  "paths": {
...
    "/api/v1/partner": {
      "get": {
        "tags": [
          "Partners"
        ],
        "summary": "Get all partners",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Partners retrieved successfully"
          }
        }
      },
      "post": {
        "tags": [
          "Partners"
        ],
        "summary": "Create new partner",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "owner_name",
                  "phone_number"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nama partner (3-50 karakter)",
                    "minLength": 3,
                    "maxLength": 50,
                    "example": "Partner Coffee Shop"
                  },
                  "owner_name": {
                    "type": "string",
                    "description": "Nama pemilik (3-50 karakter)",
                    "minLength": 3,
                    "maxLength": 50,
                    "example": "John Doe"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "Nomor telepon (format: 0xxx atau +62xxx)",
                    "pattern": "^(0|(\\+62))[0-9]{8,13}$",
                    "example": "081234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Partner created successfully"
          }
        }
      }
    },
    "/api/v1/partner/{id}": {
      "get": {
        "tags": [
          "Partners"
        ],
        "summary": "Get partner by ID",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partner retrieved successfully"
          }
        }
      },
      "put": {
        "tags": [
          "Partners"
        ],
        "summary": "Update partner",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "owner_name": {
                    "type": "string"
                  },
                  "phone_number": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Partner updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Partners"
        ],
        "summary": "Delete partner",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partner deleted successfully"
          }
        }
      }
    },
...
  }
}